segstream/pyproject.toml
[tool.poetry]
name = "segstream-crud"
version = "0.1.0"
description = "Segstream"
authors = ["Jay Ginete <jginete@innovuze.com>"]
readme = "README.md"
packages = [{include = "segstream_crud"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.1.3"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
drf-yasg = "^1.21.4"
django-cors-headers = "^3.13.0"
drf-api-logger = "^1.1.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
